web/site/requirements.txt
Django==1.8.6
Pillow==3.1.1
argparse==1.2.1
distribute==0.6.24
gunicorn==19.3.0
psycopg2==2.6.1
sorl-thumbnail==12.3
wsgiref==0.1.2
